Fix for password callback.
